requirements.txt
Django==3.1.7
djangorestframework==3.12.4
django_extensions==3.1.1
yahoo-fin==0.8.8
yfinance==0.1.59
APScheduler==3.7.0
